feat: Update ChatInput style, Add FileInputIconButton, ChatToolsDisplay
